requirements/base.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asyncio==3.4.3
attrs==23.2.0
base58==2.1.1
biom-format==2.1.16
biopython==1.84
bitarray==2.9.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
cryptography==42.0.8
cutadapt==4.9
cytoolz==0.12.3
Django==5.0.6
django-cleanup==8.1.0
django-extensions==3.2.3
django-model-utils==4.5.1
dnaio==1.2.1
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isal==1.6.1
jsonalias==0.1.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lexicon==2.0.1
lru-dict==1.2.0
magic-filter==1.0.12
Metaplex==1.1.0
mnemonic==0.21
multidict==6.0.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parsimonious==0.10.0
pillow==10.4.0
protobuf==5.27.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.2
pydantic_core==2.18.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rlp==4.0.1
rpds-py==0.18.1
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
solana==0.34.0
solders==0.21.0
sqlparse==0.5.0
toolz==0.12.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.3
wait-for-it==2.2.2
web3==6.19.0
websockets==11.0.3
xopen==2.0.2
yarl==1.9.4
zlib-ng==0.4.3
